+++ snakeAI/ppo.h
@@ -9,7 +9,7 @@
 #include <ctime>
 #include <cstdlib>
 #include "bpnn.h"
-#include "policyGradient.h"
+#include "dpg.h"
 namespace ML {
 
 class PPO
@@ -19,19 +19,19 @@ public:
     ~PPO(){}
     void CreateNet(int stateDim, int hiddenDim, int hiddenLayerNum, int actionDim,
                    int maxMemorySize = 1024, int replaceTargetIter = 256, int batchSize = 64);
-    int GreedyAction(std::vector<double>& state);
-    int Action(std::vector<double>& state);
+    int GreedyAction(std::vector<float>& state);
+    int Action(std::vector<float>& state);
     void Perceive(std::vector<Step>& trajectory);
     void ExperienceReplay(std::vector<Step>& trajectory);
-    void Learn(int optType = OPT_RMSPROP, double learningRate = 0.001);
+    void Learn(int optType = OPT_RMSPROP, float learningRate = 0.001);
     void Save(const std::string &actorPara, const std::string &criticPara);
     void Load(const std::string &actorPara, const std::string &criticPara);
     int stateDim;
     int actionDim;
-    double gamma;
-    double exploringRate;
-    double epsilon;
-    double learningRate;
+    float gamma;
+    float exploringRate;
+    float epsilon;
+    float learningRate;
     int maxMemorySize;
     int replaceTargetIter;
     int batchSize;
